WIP: rasterizer: implement

+++ common/types/types.odin
@@ -1,5 +1,5 @@
 package types
 
-Vector2 :: [2]int
+Vector2 :: [2]f32
 Vector3 :: [3]f32
 Color   :: [3]f32
